+++ eurorack/src/controls.rs
@@ -180,9 +180,6 @@ impl Controls {
         self.reconcile_detune();
         self.reconcile_scale_root();
         self.reconcile_scale_mode();
-
-        // XXX: Temporary for testing
-        self.reconcile_amplitude();
     }
 
     fn reconcile_note(&mut self) {
@@ -272,29 +269,13 @@ impl Controls {
         }
         let pot = self.last_scale_mode_pot_reading;
 
-        // XXX: CV control is disable, so it can be used as VCA for testing
-        let cv = 0.0;
-        // let cv = if self.cv3.connected() {
-        //     self.cv3.value() * 2.0 - 1.0
-        // } else {
-        //     0.0
-        // };
+        let cv = if self.cv3.connected() {
+            self.cv3.value() * 2.0 - 1.0
+        } else {
+            0.0
+        };
 
         self.parameters.scale_mode = cv + pot;
-    }
-
-    // XXX: Temporary for testing
-    fn reconcile_amplitude(&mut self) {
-        self.parameters.amplitude = if self.cv3.connected() {
-            self.cv3.value()
-        } else {
-            1.0
-        };
-    }
-
-    // XXX: Temporary for testing
-    pub fn amplitude(&self) -> f32 {
-        self.parameters.amplitude
     }
 }